refactor org-mode output to its own transpiler module
